refactor: PostResponse 와 CommentResponse 분리
